change size float to int

+++ models/netdiskcap.go
@@ -3,11 +3,11 @@ package models
 import "github.com/jinzhu/gorm"
 
 type NetdiskCap struct {
-	UserID   string  `json:"userid" gorm:"primary_key;column:userid;COMMENT:'用户标识'"`
-	Capacity float64 `json:"capacity" gorm:"COMMENT:'网盘容量,单位KB'"`
+	UserID   string `json:"userid" gorm:"primary_key;column:userid;COMMENT:'用户标识'"`
+	Capacity int    `json:"capacity" gorm:"COMMENT:'网盘容量,单位KB'"`
 }
 
-func ModNetdiskCap(userid string, cap float64) error {
+func ModNetdiskCap(userid string, cap int) error {
 	nc := NetdiskCap{userid, cap}
 	if err := db.Save(&nc).Error; err != nil {
 		return err
@@ -15,7 +15,7 @@ func ModNetdiskCap(userid string, cap float64) error {
 	return nil
 }
 
-func GetNetdiskSpareCap(userid string) (float64, error) {
+func GetNetdiskSpareCap(userid string) (int, error) {
 	var nc NetdiskCap
 	err := db.First(&nc, "userid=?", userid).Error
 	if err != nil && err != gorm.ErrRecordNotFound {
